+++ repository/book_repository.go
@@ -3,7 +3,6 @@ package repository
 import (
 	"context"
 	sqlc "readly/db/sqlc"
-	"readly/entity"
 	"strings"
 	"time"
 )
@@ -41,39 +40,13 @@ func (r *BookRepositoryImpl) CreateAuthor(ctx context.Context, name string) (*st
 	return &author.Name, nil
 }
 
-type CreateBookResponse struct {
-	ID            int64
-	Title         string
-	Description   *string
-	CoverImageURL *string
-	URL           *string
-	Author        *string
-	Publisher     *string
-	PublishDate   *entity.Date
-	ISBN          *string
-}
-
-func newCreateResponse(b sqlc.Book) *CreateBookResponse {
-	return &CreateBookResponse{
-		ID:            b.ID,
-		Title:         b.Title,
-		Description:   nilString(b.Description),
-		CoverImageURL: nilString(b.CoverImageUrl),
-		URL:           nilString(b.Url),
-		Author:        nilString(b.AuthorName),
-		Publisher:     nilString(b.PublisherName),
-		PublishDate:   entity.NewDateEntityFromNullTime(b.PublishedDate),
-		ISBN:          nilString(b.Isbn),
-	}
-}
-
 func (r *BookRepositoryImpl) CreateBook(ctx context.Context, req CreateBookRequest) (*CreateBookResponse, error) {
 	p := req.toSQLC()
 	b, err := r.querier.CreateBook(ctx, p)
 	if err != nil {
 		return nil, err
 	}
-	return newCreateResponse(b), nil
+	return newCreateBookResponseFromSQLC(b), nil
 }
 
 type CreateBookGenreRequest struct {
